refactor: improve codebase
- rename assert to assert usage to avoid name collision with node's assert

- add readme

- use `/src` folder logic
