+++ source-code/website/src/services/assert-usage/index.ts
@@ -1,0 +1,8 @@
+/**
+ * --------------------------------
+ * This index file exports code that is supposed
+ * to be used outside of this directory.
+ * --------------------------------
+ */
+
+export { assertUsage } from "./src/implementation.js";
