+++ source-code/website/src/services/assert-usage/src/implementation.ts
@@ -1,11 +1,4 @@
 /**
- * --------------------------------
- * This index file exports code that is supposed
- * to be used outside of this directory.
- * --------------------------------
- */
-
-/**
  * Asserts a condition in development environment.
  *
  * Use `assertUsage` to verify the correct usage of code by other developers.
