+++ source-code/website/src/services/markdown/src/tags/Video.tsx
@@ -1,4 +1,4 @@
-import { assertUsage } from "@src/services/assert/index.js";
+import { assertUsage } from "@src/services/assert-usage/index.js";
 
 export function Video(props: { src: string }) {
   assertUsage(
